Do wiki clash calculation in same thread
