+++ packages/greenbox-web/package.json
@@ -22,7 +22,6 @@
     "@reduxjs/toolkit": "^2.5.0",
     "@urql/core": "^5.1.0",
     "color": "^4.2.3",
-    "comlink": "^4.4.2",
     "data-of-loathing": "^1.4.0",
     "date-fns": "^4.1.0",
     "framer-motion": "^11.18.2",
